Create TransactionViewer as a base class for AddTransactionVC and EditTransactionVC
