+++ MoneyM/Scenes/EditTransaction/EditTransactionViewController.swift
@@ -16,17 +16,7 @@ protocol EditTransactionDisplayLogic: AnyObject {
 	func displayEditTransaction(_ viewModel: EditTransactionModels.EditTransaction.ViewModel)
 }
 
-class EditTransactionViewController: UIViewController {
-
-	@IBOutlet weak var amountTextField: UITextField!
-
-	@IBOutlet weak var datePickerView: UIDatePicker!
-
-	@IBOutlet weak var choiceButton: ButtonChoiceView!
-
-	@IBOutlet weak var noteTextField: UITextField!
-
-	@IBOutlet weak var selectCategoryButton: UIButton!
+class EditTransactionViewController: TransactionViewerViewController {
 
 	var interactor: EditTransactionBusinessLogic?
 
@@ -36,16 +26,11 @@ class EditTransactionViewController: UIViewController {
 
 	var delegate: EditTransactionDelegate?
 
-	private var selectedCategory: CategoryModel?
-
     override func viewDidLoad() {
         super.viewDidLoad()
 
 		setup()
-		configureAmountTextField()
-		configureNoteTextField()
-		configureChoiceButton()
-		configureFont()
+		configureConfirmButton()
 
 		if let unwrapTransaction = transaction {
 			let request = EditTransactionModels.LoadTransaction.Request(transaction: unwrapTransaction)
@@ -53,6 +38,10 @@ class EditTransactionViewController: UIViewController {
 		}
 
     }
+	
+	private func configureConfirmButton() {
+		confirmButton.setTitle(NSLocalizedString("edit.title", comment: ""), for: .normal)
+	}
 
 	private func setup() {
 		let viewController = self
@@ -66,56 +55,8 @@ class EditTransactionViewController: UIViewController {
 		interactor.presenter = presenter
 		presenter.viewController = viewController
 	}
-
-	override func viewDidAppear(_ animated: Bool) {
-		amountTextField.becomeFirstResponder()
-	}
-
-	private func configureAmountTextField() {
-		amountTextField.layer.cornerRadius = 12
-		amountTextField.layer.sublayerTransform = CATransform3DMakeTranslation(12, 0, 0)
-		amountTextField.textColor = .systemRed
-	}
-
-	private func configureNoteTextField() {
-		noteTextField.layer.cornerRadius = 12
-		noteTextField.layer.sublayerTransform = CATransform3DMakeTranslation(12, 0, 0)
-	}
-
-	private func configureChoiceButton() {
-		choiceButton.setButtonTitle(NSLocalizedString("expense.title", comment: ""), button: .first)
-		choiceButton.setButtonTitle(NSLocalizedString("income.title", comment: ""), button: .second)
-		choiceButton.delegate = self
-	}
-
-	private func configureFont() {
-		let font = CustomFonts()
-
-		let amountTextFieldPointSize: CGFloat = amountTextField.font!.pointSize
-		amountTextField.font = font.roundedFont(amountTextFieldPointSize, .bold)
-
-		let noteTextFieldPointSize: CGFloat = noteTextField.font!.pointSize
-		noteTextField.font = font.roundedFont(noteTextFieldPointSize, .semibold)
-
-		choiceButton.setButtonFont(font.roundedFont(18, .medium))
-
-		selectCategoryButton.titleLabel?.font = font.roundedFont(18, .regular)
-	}
-
-	private func resetCategory() {
-		selectedCategory = CategoriesManager.shared.defaultCategory
-		selectCategoryButton.setTitle(NSLocalizedString("select_category.title", comment: ""), for: .normal)
-	}
-
-	private func showAlertMessage(title: String, message: String) {
-		let alertController = UIAlertController(title: title, message: message, preferredStyle: .alert)
-		let alertAction = UIAlertAction(title: NSLocalizedString("ok.title", comment: ""), style: .default)
-		alertController.addAction(alertAction)
-
-		present(alertController, animated: true)
-	}
-
-	@IBAction func editButtonClicked(_ sender: Any) {
+	
+	override func confirmButtonClicked(_ sender: Any) {
 
 		let mode: TransactionModel.Mode = switch choiceButton.selectedButton {
 		case .first:
@@ -131,22 +72,8 @@ class EditTransactionViewController: UIViewController {
 		interactor?.editTransaction(request)
 	}
 
-	@IBAction func selectCategoryButtonClicked(_ sender: Any) {
+	override func selectCategoryButtonClicked(_ sender: Any) {
 		router?.routeToSelectCategory()
-	}
-
-	@IBAction func cancelButtonClicked(_ sender: Any) {
-		dismiss(animated: true)
-	}
-
-	@IBAction func amountTextFieldChanged(_ sender: Any) {
-		if let text = amountTextField.text {
-			// Remove spaces from current string "1 000" -> "1000"
-			let numberStr = text.components(separatedBy: .whitespaces).joined()
-			let number = Int(numberStr)
-			let separatorNumber = number?.thousandSeparator
-			amountTextField.text = separatorNumber
-		}
 	}
 
 }
@@ -184,26 +111,5 @@ extension EditTransactionViewController: EditTransactionDisplayLogic {
 			dismiss(animated: true)
 		}
 
-	}
-}
-
-// MARK: Select category delegate
-extension EditTransactionViewController: SelectCategoryViewControllerDelegate {
-	func selectButtonClicked(category: CategoryModel?) {
-		selectedCategory = category
-		selectCategoryButton.setTitle(category?.title, for: .normal)
-	}
-}
-
-// MARK: Button choice delegate
-extension EditTransactionViewController: ButtonChoiceDelegate {
-	func buttonClicked(button: ButtonChoiceView.Buttons) {
-		switch button {
-		case .first: // Expense button
-			amountTextField.textColor = .systemRed
-		case .second: // Income button
-			amountTextField.textColor = .systemGreen
-		}
-		resetCategory()
 	}
 }
